requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
bcrypt==4.3.0
boto3==1.38.45
botocore==1.38.45
certifi==2025.6.15
cfn-flip==1.3.0
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
durationpy==0.10
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.13
greenlet==3.2.3
h11==0.16.0
hjson==3.1.0
httptools==0.6.4
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
kappa==0.6.0
Mako==1.3.10
mangum==0.19.0
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
placebo==0.9.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
s3transfer==0.13.0
setuptools==79.0.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.1
troposphere==4.9.3
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
zappa==0.60.2
